team/chan/dag-hs-mapping-dps-session/requirements.txt
polars
google-cloud-bigquery
tqdm